requirements.txt
ansiwrap==0.8.4
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
entrypoints==0.4
fastjsonschema==2.16.2
idna==3.4
importlib-resources==5.10.0
jsonschema==4.17.1
jupyter_client==7.4.7
jupyter_core==5.0.0
nbclient==0.7.0
nbformat==5.7.0
nest-asyncio==1.5.6
papermill==2.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
pyrsistent==0.19.2
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
tenacity==8.1.0
textwrap3==0.9.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
urllib3==1.26.13
zipp==3.10.0
mlflow
fastapi
uvicorn
pystan==2.19.1.1 
fbprophet
git+ssh://git@git.mulk.net/artificial-intelligence/db_management_tool.git
seaborn
plotly
statsmodels
sklearn
ipykernel
watchfiles